[Refactor] Node data structures
