+++ engine/model_loading/public/model_loading.hpp
@@ -2,7 +2,6 @@
 
 #include "components/animation_channel_component.hpp"
 #include "cpu_resources.hpp"
-#include "include_fastgltf.hpp"
 #include "thread_pool.hpp"
 
 #include <vector>
